remove leading slash in traefik/alias example, fixes #407
When storing the alias with a leading slash the behavior described in #407 occurs. The correct way is to store the value without leading slash, resulting in correct behavior. This PR fixes the documentation so other users don't run into the same problem when they try out the example or create their own following the documentation.