fix flakiness of Prometheus integration test
The order in which the Servers are setup is not guaranteed. Therefore it
is not guaranteed that two retries MUST happen. It is only guaranteed
that at least one retry MUST happen. This is because in the integration
test we have 2 Servers and 2 Requests. In a uniform rotation through the
load balancers it MUST hit the invalid Server at least once.

Added the weight configuration setting to make it more explicit.