[kubernetes] Watch configured namespaces only.
We previously watched all namespaces, which was causing issues for users
that want to apply restrictive RBAC rules. Hence, we change the logic to
hook up one informer per namespace and kind we observe.

Note that this moves the namespace filtering into the client/informers
(where it properly belongs). A nice side effect of this is that we stop
receiving tons of events from unrelated namespaces (most notably
kube-system), which has been a source of confusion for users in the
past. We also don't need the extra event notifications as "retry
signals" in case of ephemeral API failures anymore since we can now
safely rely on a fairly low sync period for recovery purposes. This is
preferable to getting hammered with event updates at sub-second
intervals, only to have Traefik's back-off sleep of 2 seconds kicking in
very often.

A few additional improvements that made it in:

- Wait for the store synchronization to happen during WatchAll instead
  of when we process events. The wait needs to happen exactly once.
- Do not close the event channel until all controllers have stopped
  processing. This was a pending race condition that could yield to
  crash.
- Remove one superfluous event/watch channel (name changes depending on
  which method you're looking at).
- Remove the namespace-related unit tests. (Must be replaced by
  to-be-written integration tests.)
- Group methods logically.
- Improve GoDocs.