[kubernetes] Use default frontend priority of zero.
Conventionally all providers should return a priority of zero if no
override is given. The server package will then compute the priority
based on the specified rules (i.e., path and host components).

The Kubernetes provider diverges and returns the path length as
default value explicitly. This change removes the custom behavior.

As a side refactoring, we change the tests to stop comparing JSON values
which are prone to subtle but non-visible differences (namely, nil
and empty properties are not reflected). Instead, we use assert.Equal
which also does a better job at showing a readable diff.