extract metrics implementation to own package
This commit is just refactoring and should provide the same
functionality as before. It extracts the metrics implementation into its
own package and makes metrics therefore extendable to be used outside of
the middleware package in the future.

Another design goal is to simplify and align the initialisation of the Metrics
implementations. Each Metric provider should only be initialised once
and reused when configuration reloads happen. This commit accomplishes
this.