grep to-be-pulled-images directly to avoid newline issue.
The previous implementation of the `pull-images` is susceptible to
a hideous formatting bug: If one of the docker-compose files is missing
its newline, the cat-grep pipe would merge the grepped image name with
the first line of the subsequent YAML file.

For instance, we had the following file with a single whoami image missing a
newline:

BOM
whoami:
  image: emilevauge/whoami
EOM

The next compose file in order would be timeout.yml and start like this:

BOM
timeoutEndpoint:
  image: yaman/timeout
EOM

As a result, `make pull-images` would try to pull the image

emilevauge/whoamitimeoutEndpoint

and fail silently (which is another problem by the way).

Adding the missing newline fixes the problem but doesn't offer
protection against this general problem in the future. Luckily, dropping
the cat and having grep operate on the files directly prevents the issue
while also saving us one unnecessary pipe. (Note that we tell grep to
omit filenames from the result so that filtering out the second column
with awk always works regardless of the filename.)