extend metrics and rebuild prometheus exporting logic
This commit extends the metrics Registry interface and extends it's
usage in order to get more insightful metrics. It also renames
Prometheus metrics and the service label (to either backend or
entrypoint) in order to fix #1759.

In order to extend the Registry interface the MultiRegistry is used so
that single metric providers do not necessarily have to implement all of
the metrics, but only a subset. At the moment only the
Prometheus implementation supports all new metrics, but extension of the
other providers should be straightforward and can be done afterwards.

Additionally the Prometheus exporting logic was adapted. The problem
with it before was that the Prometheus client library is stateful and
samples all tracked metrics on the client. This is a problem given the
things you are tracking are dynamic. We have a metric
traefik_backend_server_up which can be used to alert given too many
servers of a backend become unhealthy. Now when Traefik learns about a
new backend it will create the concrete metric with the labels for this
backend. The problem is now that given the backend is undeployed,
Traefik will still export the metric of the old backend because it holds
the Prometheus metric state and it will only be gone after a restart of
Traefik. This leads to wrong results when you calculate e.g. how many of
your backends are healthy / all backends and thus breaks alerting. To
fix this one has to implement a custom logic for the metric sampling. I
got a confirmation for this approach on the Prometheus users google
group from Prometheus maintainers. To fix the problem I implemented a
configuration generation logic so that metrics that belong to a dynamic
configuration (e.g. backends, entrypoints) get removed  again eventually
given they belong to an old configuration and are not tracked recently.
https://groups.google.com/forum/#!topic/prometheus-users/EbumSJYsJyc

This PR also splits up the metric collection of requests, request
durations and retries into separate metrics for entrypoints and
backends. This will fix #1759. Before there was only one metric
traefik_requests_total which was tracked at the entrypoint level with
the label service="$entrypointname" and at the backend level with
service="$backendname". So at each request two traefik_requests_total
metrics were trackend and so a calculation of e.g.
sum(traefik_requests_total) would yield twice the requests that actually
happened. This is fixed by splitting up the metric into two. For other
metric providers this problem was the same and as I did not adapt those
yet, they will only count the requests now against the backends. An
extension of them should happen afterwards.