Remove health check filter from Marathon tasks.
When available, Traefik uses the last Marathon health check result to
decide whether a task should be filtered or not. This behavior is
incorrect, however, as it ignores the failure threshold denoting the
number of failed health checks it would take for Marathon to actually
consider a task as failing and trigger a restart. With Traefik not
taking the threshold into account, a task is removed from the
configuration already while Marathon may still deem it healthy.

Delaying the filtering until the moment when the failure threshold is
exceeded does not buy us much: At that point, Marathon will decide to
restart the unhealthy task anyway and cause a task state change that
will lead to a Traefik configuration update excluding the task in
question.

Consequently, we can remove the filter logic.