avoid retries when any data was written to the backend
Right now it is possible that retries happen, even when the backend
received the request already. An example scenario:

- the timeout ResponseHeaderTimeout is configured to be 3 seconds
- a request is issued against the backend and the backend starts
  processing it the backend takes overall 5 seconds to process the reuqest
- due to the timeout, go will abort the requests after 3 seconds and return a network error
- Traefik would have retried that request as it is only checking for net errors so far

To avoid this, the approch is to use httptrace to disable retries as
soon as there was anything written to the backend.

Websocket requests can't be retried at this point in time.
This is due to the fact that gorilla/websocket doesn't use the request
context and so we don't get httptrace information. Websocket clients
should however retry on their own anyway.