Allow overriding the log level in debug mode.
Enabling the debug mode caused the log level to be hard-wired to DEBUG.
We change the behavior so that an explicitly defined log level always
overrides. One primary motivation is to allow pprof-based benchmarking
without excessive DEBUG logging distorting the results.

We also remove a piece of code from SetEffectiveConfiguration which was
never effective because the log level would already be set previously.