improve Prometheus metric removal logic
Formerly metric removal was done by tracking configuration generations.
This was an easy first solution but is actually not really correct and
for setups where Traefik does not receive a lot of configuration
updates, old metrics can hang around for quite some time.

The new approach is to keep track of Traefik's current configuration and
to remove metrics that "belong" to an outdated configuration, once they
were scraped. This way we can be sure that we don't keep old metrics too
long and we can ensure that metrics for short-lived configuration still
appear when the /metrics is scraped, as we only remove them after the
scrape happened.