delete Prometheus metric on metric removal
Given a metric is removed due to it belonging to an outdated part of the
configuration, we also must delete the underlying Prometheus metric.
Otherwise, given the formerly removed metric is tracked once more (with
the same label pairs), it will start off from the value it had before
the removal.