Fix missing frontend if the last ingress is broken
A frontend would be flapping between available and unavailable if
1. The hostname/path was defined by multiple ingress objects
2. One of the ingress objects references a non-existing service

The flapping was caused by undefined ordering in a golang map: if the
last processed ingress was broken a delete was issued on the frontends
map - even if previous runs succeeded.

The new logic only adds a frontend to the map of frontends if the basic
validation passed. The frontend will thus always be available if at
least one ingress definition is not broken.