TXM-3228 [DM] Fix locking around Stats publication
- The TxM fork extends the inbuilt metrics StatsRecorder.  The sequence of
using the granular set of mutexes around the StatsRecorderwas not correct.
So ensure the BackendRequest stats is also guarded for reads and writes.
- Have not added a test as there are no existing tests for this class. The previous
issue could be exposed by calling /api/health while requests are being processed
by Traefik which will be writing to the stats structures while the health endpoint
is reading them