Do not Errorf during file watcher verification test loop.
The loop in TestProvideWithWatch is meant to be executed multiple times
until either all configuration elements are found or the timeout
expires. By calling assert.*, however, an assertion failure in a
non-terminal loop run will cause t.Error to be invoked, thereby
marking the whole test as failed even though subsequent loop runs could
still yield a positive test result overall.

We change the logic to t.Error only if the timeout hits (that is, we
fail to collect the desired configuration updates in time).