Return an error if kv store CA cert is invalid
The ca cert can be read as a file or as the ca cert value directly.
If the file path is not correct, it will be read as value and added to
the ca store.
But as the cert validity is not checked, if the file path is not
correct, this is not reported and then etcdv3 will hang and return a
timeout error instead of complaining about the invalid ca cert value.
This commit is fixing the problem by returning an error is the ca cert
is invalid.