Remove X-Forwarded-Uri and X-Forwarded-Method headers if a request comes from an untrusted ip.
Failing to do so would result in an inconsistent and insecure behaviour
for the forward authentication. If TrustForwardHeader is active in the
forward auth, the auth endpoint receives these two headers if they
have been sent from an untrusted ip. Whereas other X-Forward headers
have been removed in the request and are safe to use in the auth
endpoint.