Test and exit for jq error before domain loop
In an automated (inotify) environment, there are cases where this script gets executed before Traefik actually finishes all writes to `acme.json`. This can result in `jq` errors when it's called in the the `domain` loop. The script did not check the `jq` return code and subsequently exited with return code `0`. This commit improves that behavior.