API: expose runtime representation
Since the release of the alpha v2, none of the old endpoints (/api/providers/*)
of the API handler were really relevant anymore since the separation of providers
did not make sense anymore. As a stop-gap, the most interesting information
was accessible through the /rawdata endpoint, and it basically exposed the dynamic
configuration, as used by Traefik to initialize all the routers, services,
middlewares, etc, whenever a configuration change occurs.
However, the main flaw of this representation is it did not
take into account any errors generated during said initialization, so it did not
actually represent the state of the currently running Traefik instance.

Therefore, this change creates a new representation that exposes not only the
dynamic configuration, but also said errors, as well as the dependency relations
between services, routers, and middlewares. This new representation is served at
the same /rawdata endpoint for now.

In addition, a number of related internal changes were done while redoing the
API handler.

Finally, a status handler to report on the health of servers within a service
was hooked to each service, which means the status of servers is also available
(through the exposed services) within the representation served at /rawdata.