Update api.md
The hardest part getting started with Traefik I found was that after you go through the "Get Started" section it becomes unclear.  The intro method describes using --api CLI command to access the dashboard.  Once in the documentation that is not clearly stated "if you are using [api] the CLI argument --api needs to be removed.