Update example for letsencrypt
I tried to run the example, but the container would not stay up (on RPI3, same image as above). `kubectl logs` was reporting:

```command traefik error: failed to decode configuration from flags: field not found, node: trace```

So I have decided to remove the line and the container started. I have no idea of the possible implication  of this removal, but if it is not really needed