Replace ambiguous cli help message wording
The term `API` could be mistaken with the Traefik web interface
exposing the configuration of router, services and middlewares
(As in https://docs.traefik.io/v2.0/operations/api/).

This pr replaces this term with the word `endpoint`, which is also
used in the documentation for this feature
(https://docs.traefik.io/v2.0/operations/ping/).

Fixes: #5226