Actually send header and code during WriteHeader, if needed
Before this change, the codeCatcher relied on the fact that Write would
always be called, and hence was in charge of actually sending the headers
that WriteHeader had "cached" when it had been called.

However, in the case of e.g. a 304, since the response should not have
any body, there is no reason for the caller (i.e. another proxy or
middleware in the chain of calls) to call the codeCatcher's Write.
Therefore, neither the headers, and hence nor the the response code,
were actually forwarded to the client.

Consequently, this change moves the responsibility of actually
forwarding the headers and response code to where it belongs, i.e. in
WriteHeader.

In addition, this change adds a wrapper of the codeCatcher, so that we
can satisfy the CloseNotify interface as well.

Fixes #5385