Do not initialize list of middlewares if not needed
Otherwise, when a router refers to no middleware, we would still build
an empty (but non nil) slice of middlewares in the config.
Which means, when a (e.g. docker swarm) event sends us a config that is
in theory identical, it would still look different from the current one
because it would have a nil slice, whereas the current one would have an
empty slice. Therefore, all the handlers would be unnecessarily
rebuilt.

This was particularly visible with docker swarm, because, as opposed to
the other providers, it regularly polls to check for a config change,
instead of waiting for an event from the orchestrator or app.