make a copy of headers before logging from them
The accesslog middleware reads from the response headers to create some
log entries. However, as there might be some other middlewares or
handlers along the response chain concurrently writing on the response
headers, we might end up with a concurrent read and write on the
headers.

Therefore, this PR makes a copy of the response headers, and that copy
is the one passed along solely to the accesslog for reading.

Fixes #5825