Don't throw away valid configuration updates
The configuration reloading code takes various actions in order to limit
the number of reloads that have to be performed. One of those actions is
to check if the new configuration is the same as the current
configuration. If the new configuration is the same, then there is no
need to apply it and it can be safely ignored.

However, the check was performed at the wrong location. The location
assumed that the c.currentConfigurations variable is up to date. This
assumption does not hold given the throttling and asynchronous handling
of configuration updates. The effect of this incorrect assumption is
that a configuration update can be thrown away and leave a prior
configuration in place.

This resolves the problem by moving the check to the loadMessage method.
This is the point where the c.currentConfigurations variable is modified
and so there is no chance that the new configuration will be compared
against an incorrect version.