CloseNotifier: return pointer instead of value
In bdf4c6723fba3de8a4e01ee621c7e2b70d2b42f5, we added CloseNotify
specific versions of the responseWriter wrappers in
accesslog and metrics. But the implementer is the pointer, not the
value, and we messed up because we returned the value instead of the
pointer in the "constructor".

Also, we had forgotten to remove the CloseNotify method for the type
that is not supposed to implement it.

Therefore this PR fixes the above.