Fix panic in accesslog middleware
This patch fixes an invalid memory address or nil pointer dereference
panic which route cause is still unknown.

My guess is that it occurs when a connection is closed before the
accesslog middleware had the chance to do what it's supposed to do after
the request has came back from the other middlewares.

https://github.com/containous/traefik/issues/5922

Signed-off-by: Sylvain Rabot <s.rabot@lectra.com>