Fix panic in accesslog with recover()
At the best of my current knowledge there is no way to not be racing for
`Request.Body` with `net/http.(*conn)`.

So the only way to not panic if you happen to try to read a request body
after the client connection has been closed (rogue client) is to
use recover().

Signed-off-by: Sylvain Rabot <s.rabot@lectra.com>