do not use unbounded routinepool for requests
safe.Pool is not a pool in the usual sense as it is not bounded and
it keeps on appending new goroutines, instead of reusing a limited
number of goroutines. Moreover, even when a goroutine is done, there's
still a small footprint of it, as there is a reference to it that was
appended in a slice, and that never gets removed.

For this reason, it should not be used in places where potentially lots
of go routines are generated, such as in the mirroring service, where
there's one go routine per incoming request.

Fixes #6125