Use specified network for "container" network mode
When defining a service in the docker provider which is pointing to a
container with a networking mode of `host:other-container`, the
`traefik.docker.network` was ignored and an arbitrary IP of the
`other-container` was used as the upstream.

If that IP is not routable from the Traefik instance, this could cause
Gateway Timeouts.

This change adds steps to parse configuration from labels on the
connected container from which the service container inherits its
network.

My strawman fix prioritises the `traefik.docker.network` label on the
network-owning container, but falls back to the network specified on the
service container. Perhaps this should be flipped, but the
network-owning container is the authoritative source of which networks
it is connected to, so this makes sense to me.

NOTE: This bug is also present when `--providers.docker.network` is set,
which was also surprising to me. It is unclear to me at this time
whether this patch also fixes that (but I don't think it does), as I
don't understand how this functions yet.