Makefile: be consistent with host.docker.internal on all platforms
If one considers the matrix of 4 cases of "Host (Linux or MacOS)" X "In docker
or not", commit 0c90f6afa24ef390fec43ca654f806915e821daa added a code path that
can be revealed in only 1 out of these 4 cases (and assuming host networking is
enabled when we are in docker):
when we are in docker, and when the host is MacOS, because that is the only
circumstance where 1) we lookup host.docker.internal, and 2) it actually
resolves (it is defined by default on docker for Mac, but not on docker linux).

In that circumstance, two of the pre-existing cases in TestDockerGetIPAddress
actually get in that code path, and they were not adapted to take that into
account, i.e. they expect the IP to be 127.0.0.1 , whereas host.docker.internal
will be used, and it is set to something like 192.168.65.2. So commit
0c90f6afa24ef390fec43ca654f806915e821daa broke these tests.

Therefore, to make these tests behave like they used to, this change forces
host.docker.internal to resolve to 127.0.0.1 for the docker container that run
these tests. It is only necessary for MacOS for now, but it might be needed for
linux as well in the future (See https://github.com/moby/moby/pull/40007).