Reorder migrations for v2 minor upgrades
Currently, the order of migrations is: `2.2.2 -> 2.2.5`.  `2.0 -> 2.1`, `2.1 -> 2.2`

This order is rather confusing, and I would propose to make it follow version numbers either top down, or bottom up. 
As someone wo is upgrading from 2.0 to 2.2, I can now follow the migrations one by one without checking the whole document, if there are other changes that might be important for the version skips I try to do.