fix: fix flakyness in TestTLSClientHeaders
The condition for Traefik to be ready was based on a PathPrefix(`/`) which is
also used in the dashboard router rule. So seeing this PathPrefix(`/`) doesn't
mean the docker provider was ready. To make sure we wait for the docker configuration
the PathPrefix has been changed to `/foo` to avoid the ambiguity.