Allow overriding the service name in consul.
Consul allows to register multiple instances behind a single service
name with different tags. For instance it is possible to register the
service `backend-api` twice with the following tags:

 * ["traefik.enable=true", "dev"]
 * ["traefik.enable=true", "prod"]

Without this patch, traefik would generate a single service and consider
all those instances the same. After this patch it is possible to
reconfigure the services like this:

 * ["traefik.enable=true", "traefik.service=backend-api-dev", "dev"]
 * ["traefik.enable=true", "prod"]

and traefik would be able to distinguish those two services & route
accordingly.