Add PathExcept as a new Router Rule
Sometimes it is necessary to limit the allowed paths. Currently,
however, only a positive logic is provided to allow paths. The
possibility to exclude paths is missing so far and was requested several
times, e.g., in #4812.

One approach was #4814 which created a middleware for this purpose.
However, this pull request got stalled and the decision was made that
this should be part of the routing, not the task of a middleware.

Thus, this is another approach which adds a one-time negative logic to
the routing rules. The PathExcept rule denies the paths defined in it,
which may be specified as regular expressions.

Since this rule alone is pointless, it must be used together with
another path rule (Path or PathPrefix). This other rule allows all
subpaths, while the PathExcept rule defines exclusions for this.