ratelimiter: take into account req.RemoteAddr when checking against pool of IPs
When the ratelimiter middleware is configured with a
SourceCriterion.IPStrategy.ExcludedIPs, it effectively creates a pool of IPs
that are considered as a common source of requests, and for which a common
ratelimiting is going to apply.

To check whether an incoming request matches with a given pool of IPs, the
corresponding strategy was only considering the Forwarded header of the incoming
request. In this context, it seems like it would make more sense to also check
the RemoteAddr field of the request, in addition to the Forwarded header, which
is what this change does.

Even though the change affects the pkg/ip, it only touches the CheckerStrategy
(now renamed as PoolStrategy), which is at the moment only used by the
ratelimiter middleware.

This change also clarifies the documentation in order to avoid the
misconceptions originating from the ambiguously named ExcludedIPs
option.

Fixes #8052