Fix API error response content type
Traefik's information API uses the writeError utility function
to write HTTP responses to clients in the case of an error. The
function marshals an error message to JSON, then calls
http.Error. However, http.Error sets the Content-Type header to
"text/plain; charset=utf-8". While tests for the api package
assert that the Content-Type of API responses is
"application/json," these tests return early if a test case
doesn't include a reference to a JSON file. Since tests for
errors don't usually assert the content of a JSON file, these
tests don't catch the Content-Type mismatch.

This commit inlines the second call to http.Error in writeError,
but sets the Content-Type to application/json. It preserves the
first call to http.Error since in that case, no JSON would have
been marshaled.