Add configurable tags to influxdb metrics
This allows users to set additional tags on their traefik instances
so they can distinguish different traefik instances in the influxdb.

Most of this commit is very clean. The influx-client already supports
a map of additional tags, so it was simple to add a map to the
influxdb config type and push that into the client constructor.

However, there was a little snag - I wanted to run the metrics
unit tests in two cases - one with additional tags and one without
in order to test both the inclusion of the tags, and to check
absence of problems without additional tags. This displayed an
interesting problem: Only the first call to <RegisterInfluxDB>
in the influxdb tests had any effect, because once an influxClient
was set, it would never change. So the tests always used the config
from whatever test ran first.

I've changes this so RegisterInfluxDB always sets the influxdb client.
This should be safe, because in an actual non-test traefik run, this
method is only called once during server startup anyway.