kv: support /traefik/alias prefix from v1 again
This commit adds support for atomic configuration changes in the KV
stores by allowing to change the key-prefix from that the configuration
is retrieved.

If the traefik/alias entry exist in the KV store, it's value is used as
prefix instead of the one from p.RootKey.

The implementation is basically the same then the one for Traefik v1:
- https://github.com/traefik/traefik/pull/203
- https://doc.traefik.io/traefik/v1.7/user-guide/kv-config/#atomic-configuration-changes