integration: use VPN for integration tests (for Mac)
Our integration tests run everything by default within docker.

One reason for that (among others) is if the host is a Mac, a program running on
the host can't reach a service running in a container (roughly, because there's
a Linux VM in between the two). Which means we can't ideally have e.g. Traefik
running on the host, the go tests (typically some client code) running on the
host, and the necessary third-parties (e.g. some specific web servers) living in
containers. Which is why both Traefik and the test client code also have to live
in docker by default.

This is not only pretty cumbersome (when iterating on traefik, a docker image
has to be rebuilt everytime one wants to rerun the tests), but it also makes
running the test suites extremely slow.

To address this problem, this change adds a way to setup a VPN (namely Tailscale)
that connects the host to the containers, which allows leaving the third-parties in
containers, and running Traefik and the test client code directly on the host.