customerrors: fix bug for when custom page is large enough
When the custom page used in the response is large enough to trigger a
Flush, the wrong status code was sent (200 instead of whatever was
wanted).

This PR fixes the above bug, while simplifying the code, and making it
more efficient since the response body is now forwarded directly instead
of being buffered.