feat: Add traffic size metrics
This commit superseeds the PR from @OmarElawady by adding
`bytes_sent` and `bytes_received` metrics for all metric backend.

It refactors the access log middleware to use the new capture
middleware.

This new capture middleware extract request/response size and store it
in the request context.

As pilot is deprecated, this commit does not add support for this metric
backend.