Add `rejectStatusCode` option to `IPAllowList` middleware
If not specified, this defaults to the old behavior (returning a HTTP
403 Forbidden).

Sometimes it's preferable to return a 404 rather than a 403 when
rejecting a request. For example, you may not want to leak the fact that
a page even exists.

I see some other people have asked for this. See:
- https://www.reddit.com/r/Traefik/comments/xkcxhb/hide_forbidden_pages_from_public_view/
- https://github.com/traefik/traefik/issues/2039: It's not clear to me
  if the OP was requesting something specific to the `IPAllowList` or
  not, but there are definitely some people asking for this in the
  thread.