docs: improve errorpages examples to avoid confusion
While trying to implement the `errors` middleware, I ran into issues
understanding the documentation. Using the Dynamic File Provider,
I did not realise that writing the following wouldn't work:
```yaml
http:
  middlewares:
    error-pages:
      errors:
        status:
         - "502,504"
        service: error-pages
        query: "/{status}.html"
  services:
    error-pages:
      loadBalancer:
        - url: {{ env "ERROR_PAGES_INTERNAL_URL" }}
```
The following line started appearing in my logs:
 > ERR error="strconv.Atoi: parsing \"502,504\": invalid syntax"

After some digging through the code, I found that the comma splitting
did not occur in the same place as the dash splitting. This lead me
to conclude it is likely provider specific syntax.

Therefore it makes sense to me to provide more comprehensive provider
specific examples. I have tried to achieve this with this PR.

This would be my first contribution to Traefik, so if there's anything
I need to take care of before this PR can be reviewed or merge, please
let me know and I will take care of it ASAP, thanks!