fix: use matching or only service for routers
When the router configuration does not specify a serviceName, and a
service with the same name as the router is found, use that service.
If there is only a single service, use that service.

This allows the configuration to omit configuring the serviceName for
routers that have an identical name to the service.

Fixes: https://github.com/traefik/traefik/issues/4807
Also: https://community.traefik.io/search?q=%22too%20many%20services%22