Surface provider configuration conflicts to API
Currently, when traefik loads dynamic configurations and merges them
together it will check to see if they conflict. When they conflict, the
resultant object (typically Router, Loadbalancer, or Middleware) will
be dropped from the configuration update. This is by design.

Unfortunately, this misconfiguration is very difficult to detect in the
field as it can only be discovered by observing the *lack* of the
desired config in the API or examining the traefik daemon logs.

This change instead exposes an error field to the Router configuration
such that when populated the runtime will automatically disable the
route. This is similar to how syntactically invalid route rules are
handled today.

Related:
  - traefik/traefik#7738

There are a few other issues related to surfacing providers errors more
 effectively, but they active ones seem to focus on metric based
 detection. Which while useful, doesn't help me at the time of
 publishing an update to verify changes were valid.