Support for watching instead of polling Nomad
This change modifies the Nomad provider to allow watching for service events to trigger reconfiguration.

Backwards compatibility is maintained by defaulting to polling as usual with a configurable interval. The new `watch` option can be set to `true` to enable event watching. When watching is enabled, the provider will listen for service events from the Nomad cluster and trigger reconfiguration as needed subject to the throttling setting.

The actual code change is modelled after how the Kubernetes provider handles event throttling and consumption.

The polling mechanism that the Nomad provider currently uses leads to longer latency for updates.

The throttling code and some structure is lifted from the Kubernetes provider. There may be a more elegant way to share the code but that seemed out of scope for this PR.