Allow configuration of ACME provider http timeout
Previously, provisioning certificates via an ACME server uses a
hard-coded http client timeout of two minutes total and 30 seconds for
individual connection stages. This could lead to unexpected timeouts and
provisioning failures when the ACME server was too slow in responding.

This commit adds a new option 'ClientTimeout' to the HTTPS Provider
Static configuration to allow configuring this timeout. When set, the
provided duration is passed to the underlying *http.Client used for
communicating with the ACME server. If the option is not set (the
default) the timeout defaults to 30 seconds, matching previous behavior.