Adjust debug log message for TLS sanitization
While debugging a TLS-related issue yesterday, I initially assumed that a certificate was missing because it did not appear in the debug configuration log. After some time I realized that Traefik intentionally hides TLS fields when logging the dynamic configuration.

To avoid similar confusion in the future, I adjusted the debug log message to explicitly state that TLS data is sanitized. This is a small, non-functional change that improves clarity during debugging.

Signed-off-by: Michal K <kuritka@gmail.com>